+++ src/pages/users/signup.tsx
@@ -3,7 +3,6 @@ import Router from "next/router";
 
 import UserForm from "../../components/Users/Form";
 import Messages from "../../utils/messages";
-import { isLoggedIn } from "../../utils/auth";
 import { useCurrentUser } from "../../hooks";
 
 const SignUp = () => {
@@ -13,7 +12,7 @@ const SignUp = () => {
     if (currentUser?.isAuthenticated) Router.push("/");
   }, [currentUser]);
 
-  if (isLoggedIn(currentUser)) return <>{Messages.users.alreadyRegistered()}</>;
+  if (currentUser) return <>{Messages.users.alreadyRegistered()}</>;
   return <UserForm />;
 };