+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 # Changelog
 
+## 6/6/2021
+
+### Added
+
+- #31 Set up testing environment
+  - Unit tests for Header
+  - Contributing documentation
+  - Code of Conduct
+
 ## 4/14/2021
 
-- Added changelog
+### Added
+
+- Changelog